Refactor: Improve application structure and functionality.
This commit introduces several improvements to the React application:

- Simplified the `useEffect` hook in `App.js` for saving appointments to `localStorage`, removing redundant reads.
- Implemented unique IDs for appointments:
    - `crearCita` in `App.js` now assigns a unique ID (`Date.now().toString()`) to each new appointment.
    - `Cita` components use this unique ID as the `key` prop.
    - Deletion logic (`eliminarCita` in `App.js` and the corresponding call in `Cita.js`) now uses the unique ID, making it more reliable than index-based deletion.
- Improved code organization by separating components:
    - `Cita` component moved to `src/components/Cita.js`.
    - `Formulario` component moved to `src/components/Formulario.js`.
    - `App.js` updated to import these components.
- Added basic input validation to `Formulario.js`:
    - The `enviarCita` function now checks if all fields (`mascota`, `propietario`, `fecha`, `hora`, `sintomas`) are filled.
    - An alert is displayed if any field is empty, preventing form submission.
