+++ src/components/Cita.js
@@ -1,0 +1,20 @@
+import React from 'react';
+
+function Cita({ cita, eliminarCita, id }) {
+  return (
+    <div className="cita">
+      <p>Mascota: <span>{ cita.mascota }</span></p>
+      <p>Dueño: <span>{ cita.propietario }</span></p>
+      <p>Fecha: <span>{ cita.fecha }</span></p>
+      <p>Hora: <span>{ cita.hora }</span></p>
+      <p>Síntomas: <span>{ cita.sintomas }</span></p>
+      <button
+        className="button eliminar u-full-width"
+        onClick={ () => eliminarCita(id) }
+      >Eliminar X
+      </button>
+    </div>
+  );
+}
+
+export default Cita;
